LPD-50348 Rename "customized" sample to "advanced", in order to avoid confusion with custom data sets. Also explain what features are exercised by each sample
