+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/resources/META-INF/resources/js/index.ts
@@ -3,18 +3,18 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
+import AdvancedPropsTransformer from './AdvancedPropsTransformer';
 import ControlledFrontendDataSet from './ControlledFrontendDataSet';
 import CustomAuthorTableCell from './CustomAuthorTableCell';
 import CustomInternalViewPropsTransformer from './CustomInternalViewPropsTransformer';
-import CustomizedPropsTransformer from './CustomizedPropsTransformer';
 import GreenCheckColorTableCell from './GreenCheckColorTableCell';
 import SampleReactFrontendDataSet from './SampleReactFrontendDataSet';
 
 export {
+	AdvancedPropsTransformer,
 	ControlledFrontendDataSet,
 	CustomAuthorTableCell,
 	CustomInternalViewPropsTransformer,
-	CustomizedPropsTransformer,
 	GreenCheckColorTableCell,
 	SampleReactFrontendDataSet,
 };
