+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/resources/META-INF/resources/partials/empty.jsp
@@ -11,6 +11,7 @@
 EmptyFDSDisplayContext emptyFDSDisplayContext = new EmptyFDSDisplayContext(request);
 %>
 
+<p>Headless display tag customizing the empty state</p>
 <frontend-data-set:headless-display
 	apiURL="<%= emptyFDSDisplayContext.getAPIURL() %>"
 	emptyState="<%= emptyFDSDisplayContext.getEmptyState() %>"
